RF: Updated folder strucutre.
- wgsl folder changed.
- removed not used wgsl.
