+++ dipy/viz/meson.build
@@ -5,7 +5,6 @@ python_sources = [
   'plotting.py',
   'projections.py',
   'regtools.py',
-  'sh_billboard.py',
   'streamline.py',
   ]
 
@@ -20,4 +19,3 @@ py3.install_sources(
 subdir('horizon')
 subdir('skyline')
 subdir('tests')
-subdir('wgsl')
