+++ dipy/viz/skyline/wgsl/sh_cube_hermite_lut_compute.wgsl
