--- dipy/viz/wgsl/meson.build
@@ -1,14 +1,0 @@
-wgsl_sources = [
-  '__init__.py',
-  'billboard_sphere_render.wgsl',
-  'sh_analytical_lut_populate.wgsl',
-  'sh_billboard.wgsl',
-  'sh_cube_hermite_lut_compute.wgsl',
-  'utils.wgsl',
-]
-
-py3.install_sources(
-  wgsl_sources,
-  pure: false,
-  subdir: 'dipy/viz/wgsl'
-)
